Fix TSAN complaints: max stacksize and response stats size counters.
Start using new generic histogram code and while there, fix a bug
concerning the rcode stats (which was not passed in by the caller).

+++ pdns/mtasker.cc
@@ -399,7 +399,7 @@ template<class Key, class Val>int MTasker<Key,Val>::getTid() const
 }
 
 //! Returns the maximum stack usage so far of this MThread
-template<class Key, class Val>unsigned int MTasker<Key,Val>::getMaxStackUsage()
+template<class Key, class Val>uint64_t MTasker<Key,Val>::getMaxStackUsage()
 {
   return d_threads[d_tid].startOfStack - d_threads[d_tid].highestStackSeen;
 }
